extract transition-making logic to its own module
This separates the table module, which deals with table *construction*, and the transitions module, which deals with table *examination*.

Casualties include the missing_docs lint, which had to go as I had to make lots of stuff public, and a few re-exports in lib.
