+++ src/lib.rs
@@ -1,8 +1,12 @@
 //! Parsing Olson DB formats.
 
+#![crate_name = "zoneinfo_parse"]
+#![crate_type = "rlib"]
+#![crate_type = "dylib"]
+
 #![warn(missing_copy_implementations)]
 #![warn(missing_debug_implementations)]
-#![warn(missing_docs)]
+//#![warn(missing_docs)]
 
 #![warn(trivial_casts, trivial_numeric_casts)]
 #![warn(unused_qualifications)]
@@ -12,11 +16,7 @@ extern crate datetime;
 extern crate regex;
 #[macro_use] extern crate lazy_static;
 
-mod line;
-pub use line::Line;
-
-mod table;
-pub use table::{Table, TableBuilder, Saving, FixedTimespanSet, FixedTimespan};
-
-mod structure;
-pub use structure::*;
+pub mod line;
+pub mod table;
+pub mod transitions;
+pub mod structure;
